Added _PRNG_ prefix to macro temporaries.
A piece of code had defined and used a similar macro constant which PRNG
happily overwrote and then cleaned up. This adds a namespace prefix to
the macro temporaries that we use internally. It's not foolproof, but it's
an improvement.

Also removed the print statement which triggered in the driver when the system
random library was invoked. It made make hilariously loud and annoying.